+++ web/components/filter.tsx
@@ -1,16 +1,16 @@
 import {
   CompletionContext,
   CompletionResult,
+  FeatherProps,
+  FunctionalComponent,
   useEffect,
   useRef,
   useState,
 } from "../deps.ts";
 import { FilterOption } from "../types.ts";
-import { FunctionalComponent } from "https://esm.sh/v99/preact@10.11.3/src/index";
-import { FeatherProps } from "https://esm.sh/v99/preact-feather@4.2.1/dist/types";
 import { MiniEditor } from "./mini_editor.tsx";
-import { fuzzySearchAndSort } from "./fuse_search.ts";
-import { deepEqual } from "../../lib/json.ts";
+import { fuzzySearchAndSort } from "../fuse_search.ts";
+import { deepEqual } from "$lib/json.ts";
 
 export function FilterList({
   placeholder,
