+++ web/fuse_search.ts
@@ -1,6 +1,6 @@
 // @deno-types="https://deno.land/x/fuse@v6.4.1/dist/fuse.d.ts"
 import Fuse from "https://deno.land/x/fuse@v6.4.1/dist/fuse.esm.min.js";
-import { FilterOption } from "../types.ts";
+import { FilterOption } from "./types.ts";
 
 type FuseOption = FilterOption & {
   baseName: string;
