+++ web/sync_service.ts
@@ -4,7 +4,7 @@ import { SpaceSync, SyncStatus, SyncStatusItem } from "$common/spaces/sync.ts";
 import { sleep } from "$lib/async.ts";
 import { EventHook } from "$lib/plugos/hooks/event.ts";
 import { DataStore } from "$lib/data/datastore.ts";
-import { Space } from "./space.ts";
+import { Space } from "../common/space.ts";
 
 // Keeps the current sync snapshot
 const syncSnapshotKey = ["sync", "snapshot"];
